+++ src/copy/forms.ts
@@ -1,0 +1,4 @@
+export const forms = {
+    words: ['слово', 'слова', 'слов'],
+    days: ['день', 'дня', 'дней'],
+} as const
